Time: 35 ms (43.16%), Space: 15.5 MB (76.42%) - LeetHub

+++ 3195-separate-black-and-white-balls/3195-separate-black-and-white-balls.cpp
@@ -20,7 +20,7 @@ public:
             zeroIndex--;
         }
 
-        ans += (long long )(zeroIndex-oneIndex);
+        ans += (zeroIndex-oneIndex);
 
         oneIndex++;
         zeroIndex--;
